Add terraform s3 module
This commit adds in terraform s3 module to create a terraform.tfstate
bucket to hold all of our terraform state files for our future modules.
This is needed in order to track infrstructure changes across machines. We
use a s3 bucket because terraform states store sensitive data that would
not live well in version control.