Remove old terraform infrastructure
This commit removes old terraform infrastructure for ecr and fargate. This
is because we're choosing to pursue AWS lambda as our architecture and no
longer require the old infra files. Also update the .github subdirectory
files to point to new terraform workspace because we can't delete the
state in the old workspace.