Add tf state to be tracked by aws bucket
This commit adds in a subdirectory to track remote state that will be
used by the rest of the repo. While this repo is not tracked with any
state anywhere else, this is intentional in order to help address the
chicken egg problem where the rest of the infrastructure needs a s3
bucket to track state, but the state bucket also needs to be created.
Added README instructions to clarify when this remote state subdir
shoudl be run.